Made create_product_entry.html that can add product and show it on main.html that link to base.html, forms.py to set fields and model and set path on urls.py

+++ main/models.py
@@ -4,7 +4,7 @@ import uuid
 class Ecommerce(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     name = models.CharField(max_length=255, unique=True)
-    price = models.DecimalField(max_digits=13, decimal_places=3) # harga maks adalah dalam satuan miliar dengan 3 decimal
+    price = models.IntegerField() # harga maks adalah dalam satuan miliar dengan 3 decimal
     quantity = models.IntegerField()
     description = models.TextField()
